lazyft_pkg/requirements.txt
typer
loguru
pandas
sh
attrs
rich
pyyaml
rapidjson
pytest
python-box
numpy
python-dateutil
arrow
python-dotenv
pandas-ta
nb_black
pydantic
celery
pushbullet.py
